+++ bm/users/tests/test_models.py
@@ -3,7 +3,7 @@ from unittest import skip
 from test_plus.test import TestCase
 
 
-@skip("Don't want to test")
+# @skip("Don't want to test")
 class TestUser(TestCase):
 
     def setUp(self):
@@ -14,6 +14,8 @@ class TestUser(TestCase):
             self.user.__str__(),
             "testuser",  # This is the default username for self.make_user()
         )
+        print("user-id> ", self.user.id)
+        # find the user id number
 
-    def test_get_absolute_url(self):
-        self.assertEqual(self.user.get_absolute_url(), "/users/testuser/")
+    # def test_get_absolute_url(self):
+    #     self.assertEqual(self.user.get_absolute_url(), "/users/testuser/")