+++ bm/users/tests/test_utils.py
@@ -1,0 +1,7 @@
+
+from bm.users.models import User
+from bm.users.utils import import_class
+
+
+def test_import_class(value="bm.users.models.User"):
+    assert import_class(value) == User