+++ bm/users/utils.py
@@ -1,11 +1,11 @@
-import datetime.datetime as datetime
+import datetime as datetime
+import importlib
 
 
 def to_date_format(date, date_format):
     """converting the date object into
     given date format.
 
-    [description]
     :param date: [description]
     :type date: [datetime.datetime]
     :param date_format: [description]
@@ -13,4 +13,17 @@ def to_date_format(date, date_format):
     :returns: [description]
     :rtype: {[str]}
     """
-    return datetime.strftime(date, date_format)
+    return datetime.datetime.strftime(date, date_format)
+
+
+def import_class(value):
+    """Import the given class based on string.
+
+    :param value: [path of the class]
+    :type value: [str]
+    :returns: [class object]
+    :rtype: {[Object]}
+    """
+    value, class_name = value.rsplit(".", 1)
+    module = importlib.import_module(value)
+    return getattr(module, class_name)